Replace run blocks with shell blocks
Replaces custom code in Snakemake run blocks with shell commands that
either call scripts based on this custom code or third-party utilities
that do the same job but better. Changes include:

 - count titers with tsv-utils
 - select titers by strain names with tsv-utils
 - sanitize trees with script
 - assign clades with separate rules for HA and other segments

This change to shell blocks allows us to define conda environments for
each rule and run the whole workflow in a cluster environment instead of
running these run blocks as "localrules".