Make min date optional, enable max date for frequencies
Although min and max date are not required for any subsampling logic in
the workflow, the tip frequencies rule still expected a min date. It
also hardcoded a max date of today. This commit makes min date optional
to match the behavior for augur filter and also adds a parameter for max
date, so the user can specify their own ranges. The max date defaults to
"0D" which corresponds to "today" in the parlance of Augur's relative
dates implementation.