scripts/select_strains: check COVID-19 pandemic time interval
If the min date of a resolution lands in the time interval of the
COVID-19 pandemic when there was very littl flu data
(2020-04-01 to 2022-02-01), then use the hard-coded min date of
2020-01-01 to ensure that we have enough data to build a properly
rooted tree.

Note: In our flu builds, this can result in trees with more samples than
usual due to subsampling over a longer time interval.